authorize Role Added DeleteService-but still in comment

+++ E.Application/CQRS/User/Handler/CommandHandler/LoginUserCommandHandler.cs
@@ -6,8 +6,6 @@ using E.Application.CQRS.User.Command.Request;
 using E.Application.CQRS.User.Command.Response;
 using MediatR;
 using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json.Linq;
-using System.Data;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 
@@ -47,7 +45,7 @@ public class LoginUserCommandHandler : IRequestHandler<LoginUserCommandRequest, 
             var token = TokenService.CreateToken(authClaims, _configuration);
             var tokenString = new JwtSecurityTokenHandler().WriteToken(token); // Convert the token to a string
 
-            var loginUserCommandResponse = new LoginUserCommandResponse() { Token = tokenString, Expiration = token.ValidTo, UserRole = (int)user.UserRole };
+            var loginUserCommandResponse = new LoginUserCommandResponse() { Token = tokenString, Expiration = token.ValidTo, UserRole = user.UserRole };
 
             return new TypedResponseModel<LoginUserCommandResponse>
             {
